Attempt to fix type ambiguity (http://www.wildfiregames.com/forum/index.php?showtopic=12546)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7068 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/atlas/AtlasScript/ScriptInterface.cpp
@@ -182,7 +182,7 @@ namespace
 			if (! ret)
 				FAIL("Argument must be convertible to a string");
 			jschar* ch = JS_GetStringChars(ret);
-			out = wxString((const char*)ch, wxMBConvUTF16(), JS_GetStringLength(ret)*2);
+			out = wxString((const char*)ch, wxMBConvUTF16(), (size_t)(JS_GetStringLength(ret)*2));
 			return true;
 		}
 	};
